@@ -1,44 +1,59 @@
 {
-  "name": "medcattrainer",
-  "version": "0.1.0",
+  "name": "frontend2",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src"
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "test:unit": "vitest",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --fix",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-brands-svg-icons": "^6.1.1",
-    "@fortawesome/free-regular-svg-icons": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/vue-fontawesome": "^2.0.10",
-    "@ssthouse/vue-tree-chart": "^0.6.8",
-    "axios": "^1.7.4",
-    "bootstrap": "^4.6.0",
-    "bootstrap-vue": "^2.23.1",
-    "cache-loader": "^4.1.0",
-    "jquery": "^3.6.0",
-    "lodash": "^4.17.21",
-    "popper.js": "^1.16.1",
-    "sass": "~1.64.2",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-regular-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/vue-fontawesome": "^3.0.8",
+    "@ssthouse/vue3-tree-chart": "^0.2.6",
+    "axios": "^1.7.7",
+    "bootstrap": "^5.3.3",
+    "splitpanes": "^3.1.5",
+    "tiny-emitter": "^2.1.0",
     "v-runtime-template": "^1.10.0",
-    "vue": "^2.6.14",
-    "vue-cookie": "^1.1.4",
-    "vue-multipane": "^0.9.5",
-    "vue-router": "^3.5.3",
-    "vue-select": "^3.20.0",
-    "vue-sidebar-menu": "^4.8.1",
-    "vue-simple-context-menu": "^3.4.2",
-    "vue-splitpane": "^1.0.6"
+    "vue": "^3.5.12",
+    "vue-cookies": "^1.8.4",
+    "vue-router": "^4.4.5",
+    "vue-select": "^4.0.0-beta.6",
+    "vue-sidebar-menu": "^5.4.2",
+    "vue-simple-context-menu": "^4.1.0",
+    "vue3-runtime-template": "^1.0.2",
+    "vuetify": "^3.7.3"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^1.6.1",
-    "chai": "^4.3.0",
-    "eslint": "^8.42.0",
-    "eslint-plugin-vue": "^8.7.1",
-    "vite": "^2.9.18",
-    "vite-plugin-vue2": "1.9.0"
+    "@tsconfig/node20": "^20.1.4",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^20.17.0",
+    "@vitejs/plugin-vue": "^5.1.4",
+    "@vitest/eslint-plugin": "1.1.7",
+    "@vue/eslint-config-prettier": "^10.0.0",
+    "@vue/eslint-config-typescript": "^14.1.1",
+    "@vue/test-utils": "^2.4.6",
+    "@vue/tsconfig": "^0.5.1",
+    "eslint": "^9.13.0",
+    "eslint-plugin-vue": "^9.29.0",
+    "jsdom": "^25.0.1",
+    "npm-run-all2": "^7.0.1",
+    "prettier": "^3.3.3",
+    "sass": "1.77.6",
+    "typescript": "~5.6.0",
+    "vite": "^5.4.10",
+    "vite-plugin-vue-devtools": "^7.5.3",
+    "vitest": "^2.1.3",
+    "vue-tsc": "^2.1.6"
   }
 }
