@@ -1,7 +1,9 @@
 <template>
   <modal class="login" :closable="closable" @modal:close="$emit('login:close')">
-    <h3 slot="header">Login</h3>
-    <div slot="body">
+    <template #header>
+      <h3>Login</h3>
+    </template>
+    <template #body>
       <form @submit.prevent class="form">
         <label>Username:</label>
         <input v-model="uname" class="form-control" id="uname">
@@ -11,10 +13,10 @@
       </form>
       <span v-if="failed" class="text-danger">Username and/or password incorrect</span>
       <span v-if="failedAdminStatusCheck" class="text-danger">Cannot determine admin status of username</span>
-    </div>
-    <div slot="footer">
+    </template>
+    <template #footer>
       <button class="login-submit btn btn-primary" @click="login()">Login</button>
-    </div>
+    </template>
   </modal>
 </template>
 
@@ -33,6 +35,9 @@ export default {
   props: {
     closable: Boolean
   },
+  emits: [
+    'login:close'
+  ],
   components: {
     Modal
   },
@@ -53,22 +58,22 @@ export default {
       }
 
       instance.post('/api/api-token-auth/', payload, {}).then(resp => {
-        this.$cookie.set('api-token', resp.data.token, { expires: 7 })
-        this.$cookie.set('username', this.uname)
-        this.$http.defaults.headers.common['Authorization'] = `Token ${this.$cookie.get('api-token')}`
+        this.$cookies.set('api-token', resp.data.token, { expires: 7 })
+        this.$cookies.set('username', this.uname)
+        this.$http.defaults.headers.common['Authorization'] = `Token ${this.$cookies.get('api-token')}`
         window.removeEventListener('keyup', this.keyup)
         this.$http.get(`/api/users/?username=${this.uname}`).then(resp => {
           if (resp.data.results.length > 0) {
             const adminState = resp.data.results[0].is_staff || resp.data.results[0].is_superuser
             const userId = resp.data.results[0].id
-            this.$cookie.set('admin', adminState)
-            this.$cookie.set('user-id', userId)
+            this.$cookies.set('admin', adminState)
+            this.$cookies.set('user-id', userId)
             EventBus.$emit('login:success')
           } else {
             this.failedAdminStatusCheck = true
             setTimeout(() => {
-              this.$cookie.set('admin', false)
-              this.$cookie.delete('user-id')
+              this.$cookies.set('admin', false)
+              this.$cookies.remove('user-id')
               EventBus.$emit('login:success')
             }, 1000)
           }
