@@ -1,41 +1,46 @@
 <template>
   <div class="full-height metrics-reports-table">
-    <b-table id="metrics-jobs-table" hover small selectable
-             :items="reports.items"
-             :fields="reports.fields"
-             :select-mode="'single'"
-             @row-selected="loadMetrics">
-      <template #cell(projects)="data">
-        <v-runtime-template :template="data.value"></v-runtime-template>
+    <v-data-table :items="reports.items"
+                  :headers="reports.headers"
+                  :hover="true"
+                  @click:row="loadMetrics">
+      <template #item.projects="{ item }" >
+        <div @click.stop>
+          <v-runtime-template :template="projectsFormatter(item.projects)"></v-runtime-template>
+        </div>
       </template>
 
-      <template #cell(status)="data">
-        <span v-if="data.item.status == 'pending'">Pending
+      <template #item.status="{ item }">
+        <span v-if="item.status == 'pending'">Pending
           <font-awesome-icon icon="fa-regular fa-clock" class="status-icon"></font-awesome-icon>
         </span>
-        <span v-if="data.item.status == 'running'">Running
+        <span v-if="item.status == 'running'">Running
           <font-awesome-icon icon="spinner" spin class="status-icon"></font-awesome-icon>
         </span>
-        <span v-if="data.item.status == 'complete'">Complete
+        <span v-if="item.status == 'complete'">Complete
           <font-awesome-icon icon="check" class="status-icon success"></font-awesome-icon>
         </span>
       </template>
-      <template #cell(cleanup)="data">
-        <button class="btn btn-outline-danger" @click="confDeleteReportModal = data.item"
-                :disabled="data.item.status === 'pending' || data.item.status === 'running'">
-          <font-awesome-icon icon="times"></font-awesome-icon>
-        </button>
+      <template #item.cleanup="{ item }">
+        <div @click.stop>
+          <button class="btn btn-outline-danger" @click="confDeleteReportModal = item"
+                  :disabled="item.status === 'pending' || item.status === 'running'">
+              <font-awesome-icon icon="times"></font-awesome-icon>
+          </button>
+        </div>
       </template>
-    </b-table>
-    <b-overlay :show="loadingReports">
-      <template #overlay>
-        <b-spinner :variant="'primary'"></b-spinner>
-        <span class="overlay-message">Loading Metrics Reports...</span>
-      </template>
-    </b-overlay>
+    </v-data-table>
+
+    <v-overlay :model-overlay="loadingReports">
+      <v-progress-circular indeterminate color="'primary'"></v-progress-circular>
+      <span class="overlay-message">Loading Metrics Reports...</span>
+    </v-overlay>
+
     <modal v-if="confDeleteReportModal" :closable="true" @modal:close="confDeleteReportModal = null">
-      <h3 slot="header">Confirm Delete Metrics Report</h3>
-      <div slot="body">
+      <template #header>
+        <h3>Confirm Delete Metrics Report</h3>
+      </template>
+      <template #body>
         <div v-if="confDeleteReportModal.status == 'running'">
           Metrics report still running, this will cancel the running metrics report running job
         </div>
@@ -43,16 +48,17 @@
           Confirm complete metrics report deletion. To regenerate the report reselect the projects on the
           home screen and select metrics.
         </div>
-      </div>
-      <div slot="footer">
+      </template>
+      <template #footer>
         <button class="btn btn-danger" @click="confRemoval">Confirm</button>
-      </div>
+      </template>
     </modal>
+
   </div>
 </template>
 
 <script>
-import VRuntimeTemplate from 'v-runtime-template'
+import VRuntimeTemplate from 'vue3-runtime-template'
 import Modal from '@/components/common/Modal.vue'
 
 export default {
@@ -67,14 +73,14 @@ export default {
       projects: {},
       reports: {
         items: [],
-        fields: [
-          { key: 'report_id', label: 'ID' },
-          { key: 'report_name', label: 'Report Name' },
-          { key: 'created_user', label: 'Created User' },
-          { key: 'create_time', label: 'Create Time' },
-          { key: 'projects', label: 'Projects', formatter: this.projectsFormatter },
-          { key: 'status', label: 'Status' },
-          { key: 'cleanup', label: 'Remove' }
+        headers: [
+          { value: 'report_id', title: 'ID' },
+          { value: 'report_name', title: 'Report Name' },
+          { value: 'created_user', title: 'Created User' },
+          { value: 'create_time', title: 'Create Time' },
+          { value: 'projects', title: 'Projects' },
+          { value: 'status', title: 'Status' },
+          { value: 'cleanup', title: 'Remove' }
         ]
       }
     }
@@ -107,11 +113,11 @@ export default {
         })
       })
     },
-    loadMetrics (rows) {
+    loadMetrics (_, { item }) {
       this.$router.push({
         name: 'metrics',
         params: {
-          reportId: rows[0].report_id
+          reportId: item.report_id
         }
       })
     },
